+++ test/widgets/ac_calendar_widget_test.dart
@@ -22,7 +22,7 @@ void main() {
   }) =>
       MaterialApp(
         theme: ThemeData(extensions: [
-          ACCalendarThemeExtension(data: theme ?? ACCalendarThemeData())
+          ACCalendarThemeExtension(data: theme ?? ACLightCalendarThemeData())
         ]),
         home: Scaffold(
           body: SizedBox(